+++ src/xyz/oihana/schema/constants/traits/business/DocumentsTrait.php
@@ -13,6 +13,7 @@ use xyz\oihana\schema\constants\traits\business\documents\EcoFeeRuleTrait;
 use xyz\oihana\schema\constants\traits\business\documents\InvoiceTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentInstallmentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentScheduleTrait;
+use xyz\oihana\schema\constants\traits\business\documents\PurchaseOrderTrait;
 use xyz\oihana\schema\constants\traits\business\documents\QuoteTrait;
 use xyz\oihana\schema\constants\traits\business\documents\ReceiptTrait;
 use xyz\oihana\schema\constants\traits\business\documents\StatementEntryTrait;
@@ -39,6 +40,7 @@ trait DocumentsTrait
         InvoiceTrait              ,
         PaymentInstallmentTrait   ,
         PaymentScheduleTrait      ,
+        PurchaseOrderTrait        ,
         QuoteTrait                ,
         ReceiptTrait              ,
         StatementEntryTrait       ,